requirements.txt
django==1.8
djangorestframework
markdown
django-countries
gunicorn
PyYAML
pylint
pylint-django
flake8
coverage